feat(diaries) : 다이어리 목록 조회 구현

+++ src/main/java/org/sopt/Diary/service/DiaryService.java
@@ -1,6 +1,5 @@
 package org.sopt.Diary.service;
 
-import org.sopt.Diary.dto.res.DiariesResponse;
 import org.sopt.Diary.dto.req.DiaryReq;
 import org.sopt.Diary.dto.res.DiaryResponse;
 import org.sopt.Diary.entity.Category;
@@ -8,37 +7,29 @@ import org.sopt.Diary.entity.DiaryEntity;
 import org.sopt.Diary.repository.DiaryRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
-import org.springframework.web.client.HttpServerErrorException;
+
 import org.springframework.web.server.ResponseStatusException;
 
 import java.time.Duration;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
+
 
 @Component
 public class DiaryService {
 
     private final static int LIMIT_MINUTE = 5;
-    private final static int LIMIT_DIARY = 10;
-    private final DiaryRepository diaryRepository;
-    private final UserService userService;
 
-    public DiaryService(DiaryRepository diaryRepository, UserService userService) {
+    private final DiaryRepository diaryRepository;
+
+
+    public DiaryService(DiaryRepository diaryRepository) {
         this.diaryRepository = diaryRepository;
-        this.userService = userService;
+
     }
 
     final public void createDiary(final long userId, final DiaryReq diaryRequest) {
 
-        findByDiaryId(userId);
         validateTitle(diaryRequest.title());
-
-        int minutesSinceLastDiary = calculateMinutesSinceLastDiary();
-        if (minutesSinceLastDiary < LIMIT_MINUTE) {
-            throw new ResponseStatusException(HttpStatus.TOO_MANY_REQUESTS, "5분 뒤에 요청해주세요");
-        }
-
         diaryRepository.save(
                 new DiaryEntity(diaryRequest.title(),
                 diaryRequest.content(),
@@ -49,7 +40,9 @@ public class DiaryService {
     }
 
     public DiaryResponse getDiary(final long id) {
-        DiaryEntity diaryEntity = findByDiaryId(id);
+
+        DiaryEntity diaryEntity = findByDiaryIdPrivateFalse(id);
+
         return new DiaryResponse(diaryEntity.getDiaryId(),
                 diaryEntity.getTitle(),
                 diaryEntity.getContent(),
@@ -58,7 +51,9 @@ public class DiaryService {
     }
 
     public void patchDiary(final long userId, final long diaryId, final String content, final Category category) {
-        DiaryEntity diaryEntity= findByDiaryId(diaryId);
+
+        DiaryEntity diaryEntity = findByDiaryId(diaryId);
+
         if(diaryEntity.getUserId()!=userId){
             throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         }
@@ -70,66 +65,30 @@ public class DiaryService {
                 diaryEntity.getUserId()));
     }
 
+    public void deleteDiary(final long userId, final long diaryId) {
 
-    private List<DiariesResponse> toDiariesResponse(List<DiaryEntity> diaryEntities) {
-        List<DiariesResponse> diariesResponses = new ArrayList<>();
-        int count = 0;
+        DiaryEntity diaryEntity= findByDiaryId(diaryId);
 
-        for (DiaryEntity diary : diaryEntities) {
-            if (count >= LIMIT_DIARY) break;
-            DiariesResponse diariesResponse = new DiariesResponse(diary.getDiaryId(), diary.getTitle());
-            diariesResponses.add(diariesResponse);
-            count++;
+        if(diaryEntity.getUserId()!=userId){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         }
-        return diariesResponses;
-    }
-
-    public List<DiariesResponse> getDiaryList() {
-        List<DiaryEntity> diaryEntities = diaryRepository.findTop10ByOrderByCreatedAtDesc();
-        return toDiariesResponse(diaryEntities);
-    }
-
-    public List<DiariesResponse> getDiaryListSortByContent() {
-        List<DiaryEntity> diaryEntities = diaryRepository.findByContentLength();
-        return toDiariesResponse(diaryEntities);
-    }
-
-    public List<DiariesResponse> getDiaryListSortByCategory(Category category) {
-        List<DiaryEntity> diaryEntities = diaryRepository.findByCategory(category);
-        return toDiariesResponse(diaryEntities);
-    }
-
-
-
-
-
-    public void deleteDiary(Long id) {
-        DiaryEntity diaryEntity= findByDiaryId(id);
         diaryRepository.delete(diaryEntity);
     }
 
     public DiaryEntity findByDiaryId(final long id){
-        return diaryRepository.findByIdAndIsPrivateFalse(id)
-                .orElseThrow(() -> new HttpServerErrorException(HttpStatus.NOT_FOUND));
+        return diaryRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,"일기를 찾을 수 없습니다"));
     }
 
+    public DiaryEntity findByDiaryIdPrivateFalse(final long id){
+        return diaryRepository.findByIdAndIsPrivateFalse(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,"일기를 찾을 수 없습니다"));
+    }
 
-
-    public void validateTitle(String title){
+    public void validateTitle(final String title){
         if(diaryRepository.existsByTitle(title)){
             throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE,"중복된 제목은 불가능 합니다.");
         }
     }
 
-    private int calculateMinutesSinceLastDiary() {
-        DiaryEntity latestDiary = diaryRepository.findTop1ByOrderByCreatedAtDesc();
-
-        if (latestDiary != null) {
-            LocalDateTime now = LocalDateTime.now();
-            LocalDateTime latest = latestDiary.getCreatedAt();
-            Duration duration = Duration.between(latest, now);
-            return duration.toMinutesPart();
-        }
-        return Integer.MAX_VALUE;  // 다이어리가 없을 경우, 제한 시간이 없도록 큰 값 반환
-    }
 }
